+++ dags/wagl/k8s_wagl_nrt_filter.py
@@ -15,7 +15,7 @@ from airflow.operators.subdag_operator import SubDagOperator
 from airflow.contrib.sensors.aws_sqs_sensor import SQSSensor
 from airflow.contrib.hooks.aws_sqs_hook import SQSHook
 
-from infra.variables import WAGL_TASK_POOL
+from infra.pools import WAGL_TASK_POOL
 
 
 AWS_CONN_ID = "wagl_nrt_manual"
